Update JavaDocs and refactor into new packages
Now create jmh reports at target/jmh-reports
