+++ .gitignore
@@ -13,6 +13,3 @@ out
 .classpath
 .project
 build
-
-# reports produced from benchmarks
-jmh-benchmark-report.json
