+++ src/test/java/jenkins/benchmark/jmh/samples/JmhStateBenchmark.java
@@ -1,11 +1,14 @@
-package jenkins.jmh.benchmarks;
+package jenkins.benchmark.jmh.samples;
 
-import jenkins.jmh.JmhBenchmark;
-import jenkins.jmh.JmhBenchmarkState;
+import jenkins.benchmark.jmh.JmhBenchmark;
+import jenkins.benchmark.jmh.JmhBenchmarkState;
 import org.openjdk.jmh.annotations.Benchmark;
 
 import java.util.Objects;
 
+/**
+ * Sample benchmark without doing anything special to the Jenkins instance.
+ */
 @JmhBenchmark
 public class JmhStateBenchmark {
     public static class MyState extends JmhBenchmarkState {
